Empresa/automatizacion/UI_Automation-main/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
future==0.18.2
idna==3.3
iniconfig==1.1.1
mailosaur==7.9.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
pytest-env==0.6.2
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selene==2.0.0a40
selenium==3.141.0
six==1.16.0
tomli==2.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
webdriver-manager==3.5.0
